+++ apps/MusicLabelApp/api/src/Controller/InputRequest/BadRequestExceptionListener.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Masfernandez\MusicLabel\Infrastructure\Api\Controller\InputRequest;
+
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Event\ExceptionEvent;
+use Symfony\Component\HttpKernel\KernelEvents;
+
+final class BadRequestExceptionListener implements EventSubscriberInterface
+{
+    /**
+     * @inheritdoc
+     */
+    public static function getSubscribedEvents(): array
+    {
+        return [
+            KernelEvents::EXCEPTION => ['onKernelException', 0],
+        ];
+    }
+
+    public function onKernelException(ExceptionEvent $event): void
+    {
+        $exception = $event->getThrowable();
+        if ($exception instanceof BadRequest) {
+            $event->setResponse(
+                new JsonResponse(
+                    ['errors' => $exception->getErrors()],
+                    Response::HTTP_BAD_REQUEST,
+                )
+            );
+        }
+    }
+}