+++ apps/MusicLabelApp/api/src/Type/Catalog/DoctrineAlbumTitle.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Masfernandez\MusicLabel\Infrastructure\Api\Type\Catalog;
+
+use Doctrine\DBAL\Platforms\AbstractPlatform;
+use Masfernandez\MusicLabel\Catalog\Domain\Album\ValueObject\AlbumTitle;
+use Masfernandez\MusicLabel\Infrastructure\Api\Type\Shared\DoctrineAbstractType;
+
+final class DoctrineAlbumTitle extends DoctrineAbstractType
+{
+    private const MY_TYPE = 'AlbumTitle';
+
+    public function getName(): string
+    {
+        return self::MY_TYPE;
+    }
+
+    public function getSQLDeclaration(array $column, AbstractPlatform $platform): string
+    {
+        return $platform->getVarcharTypeDeclarationSQL($column);
+    }
+
+    protected function getNamespace(): string
+    {
+        return implode(
+            "\\",
+            array_slice(
+                explode("\\", AlbumTitle::class),
+                0,
+                -1
+            )
+        );
+    }
+}