+++ src/MusicLabel/Catalog/Domain/Model/Artist/ArtistRepository.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Masfernandez\MusicLabel\Catalog\Domain\Model\Artist;
-
-interface ArtistRepository
-{
-}