+++ src/Shared/Application/Service/ApplicationService.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Masfernandez\MusicLabel\Shared\Application\Service;
+
+interface ApplicationService
+{
+    public function execute(Request $request): ?Response;
+}