+++ src/Shared/Domain/Artist/ArtistId.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Masfernandez\MusicLabel\Shared\Domain\Artist;
+
+use Masfernandez\MusicLabel\Shared\Domain\ValueObject\UuidValueObject;
+
+final class ArtistId extends UuidValueObject
+{
+}