+++ tests/MusicLabel/Catalog/Application/Album/Get/CriteriaMother.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Masfernandez\Tests\MusicLabel\Catalog\Application\Album\Get;
 
-use Masfernandez\MusicLabel\Catalog\Application\Album\Get\Criteria;
+use Masfernandez\MusicLabel\Catalog\Application\Album\Criteria;
 use Masfernandez\Tests\Shared\Infrastructure\PhpUnit\FakerMother;
 
 class CriteriaMother
@@ -14,14 +14,12 @@ class CriteriaMother
         ?array $sort = null,
         ?int $page = null,
         ?int $size = null,
-        ?array $fields = null,
     ): Criteria {
         return new Criteria(
-            $exp ?? [], //@todo filters
-            $sort ?? [],
-            $page ?? FakerMother::random()->numberBetween(1, 10),
-            $size ?? FakerMother::random()->numberBetween(1, 10),
-            $fields ?? [],
+            exp:  $exp ?? [], //@todo filters
+            sort: $sort ?? [],
+            page: $page ?? FakerMother::random()->numberBetween(1, 10),
+            size: $size ?? FakerMother::random()->numberBetween(1, 10),
         );
     }
 }