+++ tests/MusicLabel/Catalog/Application/Album/Get/Single/AlbumSearcherTest.php
@@ -6,10 +6,11 @@ declare(strict_types=1);
 
 namespace Masfernandez\Tests\MusicLabel\Catalog\Application\Album\Get\Single;
 
-use Masfernandez\MusicLabel\Catalog\Application\Album\Get\AlbumResponse;
+use Masfernandez\MusicLabel\Catalog\Application\Album\AlbumAssembler;
+use Masfernandez\MusicLabel\Catalog\Application\Album\Get\Single\AlbumResponse;
 use Masfernandez\MusicLabel\Catalog\Application\Album\Get\Single\AlbumSearcher;
-use Masfernandez\MusicLabel\Catalog\Domain\Model\Album\AlbumRepository;
-use Masfernandez\Tests\MusicLabel\Catalog\Domain\Model\Album\AlbumMother;
+use Masfernandez\MusicLabel\Catalog\Domain\Album\AlbumRepository;
+use Masfernandez\Tests\MusicLabel\Catalog\Domain\Album\AlbumMother;
 use Mockery;
 use PHPUnit\Framework\TestCase;
 
@@ -28,7 +29,8 @@ class AlbumSearcherTest extends TestCase
         $albumRepository = Mockery::mock(AlbumRepository::class);
         $albumRepository->allows()->getById($query->id())->andReturns($album);
 
-        $actualResponse = (new AlbumSearcher($albumRepository))->execute($query);
-        self::assertEquals(new AlbumResponse($album->toArray()), $actualResponse);
+        $actualResponse   = (new AlbumSearcher($albumRepository))->execute($query);
+        $expectedResponse = new AlbumResponse(AlbumAssembler::fromEntityToResponse($album));
+        self::assertEquals($expectedResponse, $actualResponse);
     }
 }