+++ tests/MusicLabel/Catalog/Application/Album/Put/PutAlbumCommandMother.php
@@ -5,21 +5,21 @@ declare(strict_types=1);
 namespace Masfernandez\Tests\MusicLabel\Catalog\Application\Album\Put;
 
 use Masfernandez\MusicLabel\Catalog\Application\Album\Put\PutAlbumCommand;
-use Masfernandez\Tests\MusicLabel\Catalog\Domain\Model\Album\AlbumIdMother;
-use Masfernandez\Tests\MusicLabel\Catalog\Domain\Model\Album\AlbumPublishingDateMother;
-use Masfernandez\Tests\MusicLabel\Catalog\Domain\Model\Album\AlbumTitleMother;
+use Masfernandez\Tests\MusicLabel\Catalog\Domain\Album\AlbumIdMother;
+use Masfernandez\Tests\MusicLabel\Catalog\Domain\Album\AlbumReleaseDateMother;
+use Masfernandez\Tests\MusicLabel\Catalog\Domain\Album\AlbumTitleMother;
 
 class PutAlbumCommandMother
 {
     public static function create(
         ?string $id = null,
         ?string $title = null,
-        ?string $publishing_date = null
+        ?string $release_date = null,
     ): PutAlbumCommand {
         return new PutAlbumCommand(
-            $id ?? AlbumIdMother::create()->value(),
-            $title ?? AlbumTitleMother::create()->value(),
-            $publishing_date ?? AlbumPublishingDateMother::create()->value()
+            id:             $id ?? AlbumIdMother::create()->value(),
+            title:          $title ?? AlbumTitleMother::create()->value(),
+            releaseDate: $release_date ?? AlbumReleaseDateMother::create()->value(),
         );
     }
 }